Version 1.1.0
animalTrack 1.1.0 includes a rewrite of the dead_reckining function.
Generic methods for plotting and printing were added. Furthermore, the
dead_reckoning function output is now a "dr" S3 class instead of a
"navigate" class. The dead_reckoning function now support projection to
UTM zones, as well as geographic coordinates and Cartesian. This version
is stable and passes R CMD check.